Knockback is in even more
Added a Kinematic rigidbody the the flier on the test stage so that knockback works with it. Knockback now works when hit from the left or the right, or even if the enemy in question is not moving. Every enemy NEEDS a rigidbody for the knockback function to work, though.